server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
motor
pydantic-settings
email-validator
python-dotenv
